Change default font to Latin Modern (serif)
The current regulations say "Arial 11 pt is the preferred font/size... other cross-platform sans serif fonts are also acceptable." However, we learned in discussions with the library that this regulation is unenforced when there are good reasons to prefer another typeface (e.g. subject specific norms, character support, or committee preference). After consultation with people in several departments, it was decided that the serifed Computer Modern (or equivalent) fonts are acceptable for theses in the mathematical sciences. Future library specifications may be edited to reflect this.