Fix order of dedication \addtoToC and \chapter
Putting these commands in the correct order will prevent off-by-one ToC page numbering errors if students add custom frontmatter sections that do not end with a \newpage declaration. This shouldn't be the case for any actual theses, but it's nice for the Table of Contents to be correct by design rather than coincidence.